fix(gtfs-api): bump gtfs-api and gtfs-lib versions to fix memory issues
GTFS API previously only loaded a set number of feeds into its cache, which caused the feeds files
to be reconnected to over and over again.  Ultimately, this caused lots of file leaks and ground the
server to a halt.  These dependency bumps fix this issue by allowing all the feeds to be loaded into
the cache that memory will permit.