fix(FeedVersion.id): clean entire feed version ID (include date time string)
when the application runs on a server east of UTC timezone, a "+" is included in the feed version
ID.  BaseGTFSCache.cleanId clears this character from the string when writing/retreiving GTFS feeds
from local disk or S3, so there is a mismatch between datatools and GTFSCache.  Initial local disk
write of feed is handled by datatools, which needs to be switched to ApiMain.registerFeedSource to
close the loop on this.

Fixes #8