fix(jobs): use threadpool to execute jobs
Previously datatools-server kicked off an endless set of threads when asked to, e.g., fetch updated
feeds for an entire project. This work defines a lightweight and heavyweight ThreadPoolExecutor to
handle light and CPU-intensive tasks, respectively. Jobs are queued in the order they are requested,
so any massive set of jobs should still be performed off-peak to limit any impact on light user
activity.