fix(auth0): handle app_metadata and verification for new auth0 tenants
New Auth0 tenants no longer have access to app_metadata or user_metadata on the user profile by
default. A custom rule must be set up in Auth0 to attach these objects to the idToken. This commit
handles remapping those values to the previous app- and user_metadata fields and attaches the entire
profile to the request object (for sending back to the requesting user).

refs catalogueglobal/datatools-ui#207