refactor(editor-lock): reduce lock timeout and allow takeovers
The initial implementation of the editor lock timed out a session after 60 minutes of inactivity,
but this is ultimately just a burden on users waiting for the lock to open. It also restricted other
users from taking over a current session, but this proved to be an unnecessary precaution because
users with access to the same feed should understand the repercussions of booting another user
(i.e., their out-of-application coordination should prevent any kind of data loss).