feat(gtfs-transform): add arbitrary transforms for imported versions
This adds the ability to define a list of arbitrary transformations that should be applied to each
new feed version loaded for a particular feed source. The primary use case for now is limited to
adding/replacing a file (in this case a GTFS+) file into each new version.

re ibi-group/datatools-ui#544